+++ App/Controller/Api/ApiIndexController.php
@@ -8,6 +8,6 @@ class ApiIndexController extends CoreController
 {
     public function index()
     {
-        \libs\Core\Message::send(200, [], 'Api Index');
+        \app\Controller\Message::send(200, [], 'Api Index');
     }
 }