+++ App/Controller/Message.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace libs\Core;
+namespace app\Controller;
 
 use libs\Core\FileLogger;
 
@@ -11,16 +11,6 @@ class Message
      */
     const CODE_INFO = [
         200 => 'success',
-
-        411 => 'Database Connection Error',
-        412 => 'Data validation failed',
-
-        // Model Related
-        10400 => 'Model error, missing table name',
-        10401 => 'Model error, When updating or deleting, where is required.',
-        10402 => 'Please pass in the second parameter, the data to be inserted cannot be empty.',
-        10403 => 'There is associated data, delete operation is not allowed.',      // use for model
-        10404 => 'When there is a where condition, the value cannot be empty.',
     ];
 
     public static function send(int $code, array $data = [], string $message = '')