+++ src/tablelogger/Core/Logger.php
@@ -1,13 +1,10 @@
 <?php
 
-namespace libs\Core;
+namespace src\tablelogger\Core;
 
-/**
- * maybe we can delete this class
- */
-
-use PDOException;
+use libs\Core\FileLogger;
 use libs\Db\DB;
+use PDOException;
 
 class Logger
 {
@@ -23,20 +20,27 @@ class Logger
      */
     private static function log(string $action, int $userId, int $status, string $description, int $logType): void
     {
-        $data['user_id'] = $userId;
-        $data['description'] = $description;
+        if ($userId == 0) {
+            // Handling of special cases, Unable to get the id of the operating user
+            $data['user_id'] = 1;
+            $data['description'] = '!!Error, illegal operation user id, ' . $description;
+        } else {
+            $data['user_id'] = $userId;
+            $data['description'] = $description;
+        }
         $data['action'] = $action;
         $data['status'] = $status;
         $data['log_type'] = $logType;
         $data['request_url'] = $_SERVER['REQUEST_URI'];
-        $data['ip_address'] = $_SERVER['REMOTE_ADDR'];
-        $data['user_agent'] = $_SERVER['HTTP_USER_AGENT'];
-        try {
-            DB::link()->table('logs')->insert($data);
-        } catch (PDOException $e) {
-            // Log the error
-            FileLogger::error('Error operating data: ' . $e->getMessage());
+        if (!empty($_SERVER['HTTP_CLIENT_IP'])) {
+            $data['ip_address'] = $_SERVER['HTTP_CLIENT_IP'];
+        } else if (!empty($_SERVER['HTTP_X_FORWARDED_FOR'])) {
+            $data['ip_address'] = $_SERVER['HTTP_X_FORWARDED_FOR'];
+        } else {
+            $data['ip_address'] = $_SERVER['REMOTE_ADDR'];
         }
+        $data['user_agent'] = $_SERVER['HTTP_USER_AGENT'];
+        DB::link()->table('logs')->insert($data);
     }
 
     public static function loginin($userId, $status = 1, $description = '', $logType = 1, $itemId = '', $module = ''): void
@@ -63,11 +67,15 @@ class Logger
         self::log(LogAction::LOGINOUT, (int)$userId, (int)$status, $description, (int)$logType);
     }
 
-    public static function add($userId, $status = 1, $description = '', $logType = 1, $itemId = '', $module = ''): void
+    public static function add($userId, $status = 1, $description = '', $logType = 1, $itemId = '', $module = '', $data = ''): void
     {
         if (empty($description)) {
             if ($status == 1) {
-                $description = "The user has successfully created an {$module}, and the {$module} ID is {$itemId}";
+                if (empty($data)) {
+                    $description = "The user has successfully created an {$module}, and the {$module} ID is {$itemId}";
+                } else {
+                    $description = "The user has successfully created an {$module}, and the {$module} ID is {$itemId}, \nthe added data is: {$data}";
+                }
             } else {
                 $description = "User creation {$module} failed";
             }
@@ -75,20 +83,34 @@ class Logger
         self::log(LogAction::ADD, (int)$userId, (int)$status, $description, (int)$logType);
     }
 
-    public static function update($userId, $status = 1, $description = '', $logType = 1, $itemId = '', $module = ''): void
+    public static function update($userId, $status = 1, $description = '', $logType = 1, $itemId = '', $module = '', array $data = []): void
     {
         if (empty($description)) {
             if ($status == 1) {
                 if ($module == 'password') {
                     $description = "The user has changed password successfully";
                 } else {
-                    $description = "The user has successfully updated the {$module}, and the {$module} ID is {$itemId}";
+                    if (empty($data)) {
+                        $description = "The user has successfully updated the {$module}, and the {$module} ID is {$itemId}";
+                    } elseif (isset($data['oldData']) && isset($data['newData'])) {
+                        $oldData = json_encode($data['oldData']);
+                        $newData = json_encode($data['newData']);
+                        $description = "The user has successfully updated the {$module}, and the {$module} ID is {$itemId}, \nThe updated data is: {$newData}, \nThe original data is: {$oldData}";
+                    } else {
+                        $data = json_encode($data);
+                        $description = "The user has successfully updated the {$module}, and the {$module} ID is {$itemId}, \nThe updated data is: {$data}";
+                    }
                 }
             } else {
                 if ($module == 'password') {
                     $description = "The user failed to change password";
                 } else {
-                    $description = "User update {$module} failed";
+                    if (empty($data)) {
+                        $description = "User update {$module} failed, the {$module} ID is {$itemId}";
+                    } else {
+                        $data = json_encode($data);
+                        $description = "User update {$module} failed, the {$module} ID is {$itemId}, \nThe updated data is: {$data}";
+                    }
                 }
             }
         }